+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 2.1.1-beta
 - Add google analytics.
+- Add karura vesting redeem page.
+- Add claim KAR dApp entrance.
 
 # 2.1.0-beta
 - Add auction & crowdloans support.